Rename component from "verify" to "preview"
