+++ src/gui/static/src/app/components/pages/send-skycoin/send-preview/send-preview.component.scss
