+++ src/gui/static/src/app/components/pages/send-skycoin/send-preview/send-preview.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { SendVerifyComponent } from './send-verify.component';
+import { SendVerifyComponent } from './send-preview.component';
 
 describe('SendVerifyComponent', () => {
   let component: SendVerifyComponent;
