+++ src/gui/static/src/app/components/pages/send-skycoin/send-preview/send-preview.component.ts
@@ -1,14 +1,14 @@
 import { Component, EventEmitter, Input, OnDestroy, Output, ViewChild } from '@angular/core';
 import { WalletService } from '../../../../services/wallet.service';
 import { ButtonComponent } from '../../../layout/button/button.component';
-import { MatSnackBar, MatSnackBarConfig } from '@angular/material';
-import { parseResponseMessage } from '../../../../utils/errors';
+import { MatSnackBar } from '@angular/material';
+import { showSnackbarError } from '../../../../utils/errors';
 import { PreviewTransaction } from '../../../../app.datatypes';
 
 @Component({
-  selector: 'app-send-verify',
-  templateUrl: './send-verify.component.html',
-  styleUrls: ['./send-verify.component.scss'],
+  selector: 'app-send-preview',
+  templateUrl: './send-preview.component.html',
+  styleUrls: ['./send-preview.component.scss'],
 })
 export class SendVerifyComponent implements OnDestroy {
   @ViewChild('sendButton') sendButton: ButtonComponent;
@@ -45,11 +45,9 @@ export class SendVerifyComponent implements OnDestroy {
         this.onBack.emit(true);
       }, 3000);
     }, error => {
-      const errorMessage = parseResponseMessage(error['_body']);
-      const config = new MatSnackBarConfig();
-      config.duration = 300000;
-      this.snackbar.open(errorMessage, null, config);
-      this.sendButton.setError(errorMessage);
+      showSnackbarError(this.snackbar, error);
+
+      this.sendButton.setError(error);
       this.backButton.setEnabled();
     });
   }
