+++ src/gui/static/src/app/components/pages/send-skycoin/send-preview/transaction-info/transaction-info.component.html
